requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
plotly>=5.17.0
tqdm>=4.66.0

pip install -r requirements.txt